requirements.txt
fastapi[standard]==0.116.1
PyGithub==2.7.0
pydantic_settings==2.10.1
langchain[openai]==0.3.27